==> Code/06c_Socioeconomic_Qualifications.Rmd <==
---
title: "NEPHU First Nations Population Profile"
subtitle: "Socioeconomic Determinants of Health: Qualifications"
output:
  officedown::rdocx_document: 
    page_size:
      width: 8.27
      length: 11.69
    page_margins:
      bottom: 0.5
      top: 0.5
      right: 0.8
      left: 0.8
      gutter: 0
    reference_docx: Template.docx
    
knit: (function(input, ...) {
  rmarkdown::render(input, output_file = paste0("NEPHU First Nations Profile - ", "06C Socioeconomic Qualifications ", format(Sys.Date(), "%Y-%m-%d")), output_dir = "../Output/Dev") })
---

```{r setup, include = FALSE}
#
knitr::opts_chunk$set(echo    = FALSE, 
                      warning = FALSE, 
                      message = FALSE,
                      #
                      fig.align  = "center",
                      fig.width  = 7,
                      fig.height = 3.5,
                      #
                      ft.align = "center",
                      #
                      dpi = 300,
                      #
                      fig.topcaption = TRUE,
                      #
                      cache = FALSE)

if (!require("pacman")) install.packages("pacman")

pacman::p_load(here,
               tidyverse,
               rlang,
               janitor,
               readxl,
               sf,
               flextable,
               colorRamps,
               officer,
               officedown,
               extrafont,
               ggnewscale,
               patchwork)

# Uncomment and run these lines if you get any weird "missing font" errors
#extrafont::font_import()
#extrafont::loadfonts(device = "win")
```


```{r, include = FALSE}
#
source(here::here("Code", "00_Setup.R"))

source(here::here("Code", "Functions", "Data prep", "f_read_census_lga.R"))
source(here::here("Code", "Functions", "Data prep", "f_wrangle_census_lga.R"))
source(here::here("Code", "Functions", "Data prep", "f_read_census_agesex.R"))
source(here::here("Code", "Functions", "Data prep", "f_wrangle_census_agesex.R"))

source(here::here("Code", "Functions", "Graphs", "f_graph_grouped.R"))
source(here::here("Code", "Functions", "Graphs", "f_graph_lga.R"))
source(here::here("Code", "Functions", "Graphs", "f_graph_ctg_target.R"))
source(here::here("Code", "Functions", "Graphs", "f_graph_line_comparison.R"))
source(here::here("Code", "Functions", "Graphs", "f_graph_expected.R"))

source(here::here("Code", "Functions", "Tables", "f_table_grouped_lga.R"))
source(here::here("Code", "Functions", "Tables", "f_table_grouped_agesex.R"))
source(here::here("Code", "Functions", "Tables", "f_table_rate_year.R"))
source(here::here("Code", "Functions", "Tables", "f_table_expected.R"))
```

**Table of Contents**

```{r}
#
officer::block_toc(level = 2)
```

\newpage

```{r, include = FALSE}
# Numbering and styling for figures and tables
f_caption_figure <- function(title_text) {
  
  caption <- block_caption(title_text, 
                           style   = "caption", 
                           autonum = run_autonum(seq_id    = "Figure", 
                                                 pre_label = "Figure 6C-"))
  
  return(caption)
  
}

f_caption_table <- function(title_text) {
  
  caption <- block_caption(title_text, 
                           style   = "caption", 
                           autonum = run_autonum(seq_id    = "Table", 
                                                 pre_label = "Table 6C-"))
  
  return(caption)
  
}
```

### What is the average educational attainment for Aboriginal and/or Torres Strait Islander people living in the NEPHU catchment?

# Highest level of education

```{r, include = FALSE}
# ------------------------------------------------------------------------------
# Highest level of education - entire population, by LGA
# Data from ABS Census using TableBuilderPro
# ------------------------------------------------------------------------------
# Read in Census data
education_qualification <- f_read_census_lga("Aboriginal_Education_Qualification_LGA.csv",
                                             variable = "qualification") %>% 
  #
  dplyr::filter(!qualification %in% c("Not stated", "Not applicable", "Supplementary Codes")) 

# Data wrangling
education_qualification_non_aboriginal <- education_qualification %>% 
  #
  dplyr::mutate(qualification = dplyr::case_when(
    qualification == "Postgraduate Degree Level"                       ~ "Postgraduate",
    qualification == "Graduate Diploma and Graduate Certificate Level" ~ "Postgraduate",
    qualification == "Bachelor Degree Level"                           ~ "Undergraduate",
    qualification == "Advanced Diploma and Diploma Level"              ~ "Certificate and Diploma",
    qualification == "Certificate III & IV Level"                      ~ "Certificate and Diploma",
    qualification == "Certificate I & II Level"                        ~ "Certificate and Diploma",
    qualification == "Secondary Education - Years 10 and above"        ~ "Year 10-12",
    qualification == "Secondary Education - Years 9 and below"         ~ "Year 9 and below",
    TRUE ~ NA_character_)) %>% 
  #
  f_wrangle_census_lga(variable = qualification,
                       level    = "person",
                       dichot   = "no") %>% 
  #
  dplyr::mutate(postgrad_prop    = non_aboriginal_n_postgraduate / non_aboriginal_total * 100,
                undergrad_prop   = non_aboriginal_n_undergraduate / non_aboriginal_total * 100,
                certificate_prop = non_aboriginal_n_certificate_and_diploma / non_aboriginal_total * 100,
                year_10_12_prop  = non_aboriginal_n_year_10_12 / non_aboriginal_total * 100,
                year_9_prop      = non_aboriginal_n_year_9_and_below / non_aboriginal_total * 100) %>% 
  #
  dplyr::select(lga_name,
                postgrad_n = non_aboriginal_n_postgraduate,
                postgrad_prop,
                undergrad_n = non_aboriginal_n_undergraduate,
                undergrad_prop,
                certificate_n = non_aboriginal_n_certificate_and_diploma,
                certificate_prop,
                year_10_12_n = non_aboriginal_n_year_10_12,
                year_10_12_prop,
                year_9_n = non_aboriginal_n_year_9_and_below,
                year_9_prop,
                non_aboriginal_total)

education_qualification_lga <- education_qualification %>% 
  #
  dplyr::mutate(qualification = dplyr::case_when(
    qualification == "Postgraduate Degree Level"                       ~ "Postgraduate",
    qualification == "Graduate Diploma and Graduate Certificate Level" ~ "Postgraduate",
    qualification == "Bachelor Degree Level"                           ~ "Undergraduate",
    qualification == "Advanced Diploma and Diploma Level"              ~ "Certificate and Diploma",
    qualification == "Certificate III & IV Level"                      ~ "Certificate and Diploma",
    qualification == "Certificate I & II Level"                        ~ "Certificate and Diploma",
    qualification == "Secondary Education - Years 10 and above"        ~ "Year 10-12",
    qualification == "Secondary Education - Years 9 and below"         ~ "Year 9 and below",
    TRUE ~ NA_character_)) %>% 
  #
  f_wrangle_census_lga(variable = qualification,
                       level    = "person",
                       dichot   = "no") %>% 
  #
  dplyr::mutate(postgrad_prop    = aboriginal_n_postgraduate / aboriginal_total * 100,
                undergrad_prop   = aboriginal_n_undergraduate / aboriginal_total * 100,
                certificate_prop = aboriginal_n_certificate_and_diploma / aboriginal_total * 100,
                year_10_12_prop  = aboriginal_n_year_10_12 / aboriginal_total * 100,
                year_9_prop      = aboriginal_n_year_9_and_below / aboriginal_total * 100) %>% 
  #
  dplyr::select(lga_name,
                postgrad_n = aboriginal_n_postgraduate,
                postgrad_prop,
                undergrad_n = aboriginal_n_undergraduate,
                undergrad_prop,
                certificate_n = aboriginal_n_certificate_and_diploma,
                certificate_prop,
                year_10_12_n = aboriginal_n_year_10_12,
                year_10_12_prop,
                year_9_n = aboriginal_n_year_9_and_below,
                year_9_prop,
                aboriginal_total)

# ------------------------------------------------------------------------------
# Highest level of education - entire population, by age and sex
# Data from ABS Census using TableBuilderPro
# ------------------------------------------------------------------------------
# Read in Census data
education_qualification_agesex <- f_read_census_agesex("Aboriginal_Education_Qualification_AgeSex.csv",
                                                       variable = "qualification") %>% 
  #
  dplyr::filter(!qualification %in% c("Not stated", "Not applicable", "Supplementary Codes")) 

# Data wrangling
education_qualification_agesex <- education_qualification_agesex %>% 
  #
  dplyr::mutate(qualification = dplyr::case_when(
    qualification == "Postgraduate Degree Level"                       ~ "Postgraduate",
    qualification == "Graduate Diploma and Graduate Certificate Level" ~ "Postgraduate",
    qualification == "Bachelor Degree Level"                           ~ "Undergraduate",
    qualification == "Advanced Diploma and Diploma Level"              ~ "Certificate and Diploma",
    qualification == "Certificate III & IV Level"                      ~ "Certificate and Diploma",
    qualification == "Certificate I & II Level"                        ~ "Certificate and Diploma",
    qualification == "Secondary Education - Years 10 and above"        ~ "Year 10-12",
    qualification == "Secondary Education - Years 9 and below"         ~ "Year 9 and below",
    TRUE ~ NA_character_)) %>% 
  #
  f_wrangle_census_agesex(variable = qualification,
                          dichot   = "no") %>% 
  #
  dplyr::mutate(aboriginal_total = aboriginal_n_postgraduate + aboriginal_n_undergraduate + aboriginal_n_certificate_and_diploma + aboriginal_n_year_10_12 + aboriginal_n_year_9_and_below,
                #
                postgrad_prop    = aboriginal_n_postgraduate / aboriginal_total * 100,
                undergrad_prop   = aboriginal_n_undergraduate / aboriginal_total * 100,
                certificate_prop = aboriginal_n_certificate_and_diploma / aboriginal_total * 100,
                year_10_12_prop  = aboriginal_n_year_10_12 / aboriginal_total * 100,
                year_9_prop      = aboriginal_n_year_9_and_below / aboriginal_total * 100) %>% 
  #
  dplyr::mutate(across(where(is.numeric), ~replace(., is.na(.), 0)),
                across(where(is.numeric), ~replace(., is.nan(.), 0))) %>% 
  #
  dplyr::select(age_group,
                sex,
                postgrad_n = aboriginal_n_postgraduate,
                postgrad_prop,
                undergrad_n = aboriginal_n_undergraduate,
                undergrad_prop,
                certificate_n = aboriginal_n_certificate_and_diploma,
                certificate_prop,
                year_10_12_n = aboriginal_n_year_10_12,
                year_10_12_prop,
                year_9_n = aboriginal_n_year_9_and_below,
                year_9_prop,
                aboriginal_total)
```


```{r, include = FALSE}
#
results <- list()

results$anyqual <- education_qualification_lga %>% 
  dplyr::mutate(anyqual_n    = postgrad_n + undergrad_n + certificate_n,
                anyqual_prop = anyqual_n / aboriginal_total * 100) %>% 
  #
  dplyr::select(lga_name,
                aboriginal_n = anyqual_n,
                aboriginal_prop = anyqual_prop)

results$anyqual_non <- education_qualification_non_aboriginal %>% 
  dplyr::mutate(anyqual_n    = postgrad_n + undergrad_n + certificate_n,
                anyqual_prop = anyqual_n / non_aboriginal_total * 100) %>% 
  #
  dplyr::select(lga_name,
                non_aboriginal_n = anyqual_n,
                non_aboriginal_prop = anyqual_prop)

data     <- results$anyqual
data_non <- results$anyqual_non

results$nephu_n <- data %>% 
  dplyr::filter(lga_name == "NEPHU") %>% 
  pull(aboriginal_n)

results$nephu_prop <- data %>% 
  dplyr::filter(lga_name == "NEPHU") %>% 
  pull(aboriginal_prop)

results$non_prop <- data_non %>% 
  dplyr::filter(lga_name == "NEPHU") %>% 
  pull(non_aboriginal_prop)

results$comp_non <- dplyr::case_when(
  results$nephu_prop < results$non_prop * 0.9 ~ "lower than for",
  results$nephu_prop > results$non_prop * 1.1 ~ "higher than for",
  TRUE ~ "similar to")

results$rank_n <- data %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  dplyr::arrange(desc(aboriginal_n))

results$rank_prop <- data %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  dplyr::arrange(desc(aboriginal_prop))
```

::: {custom-style="Dotpt"}
- Overall, `r round(results$nephu_prop, digits = 0)`% (n = `r formatC(results$nephu_n, big.mark = ",")`) of people living in the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census have a tertiary qualification (Certificate III or above). This is `r results$comp_non` non-Indigenous people living in the NEPHU catchment (`r round(results$non_prop, digits = 0)`%).

- The number of people living in the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census who have a tertiary qualification (Certificate III or above) ranged from 604 (Whittlesea LGA) to 87 (Manningham LGA).

- The percentage of people living in the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census who have a tertiary qualification (Certificate III or above) ranged from 61% (Yarra LGA) to 41% (Hume LGA).

- Younger people were more likely to have a tertiary qualification.

- Females were more likely to have a university degree, and males were more likely to have a high school level education.
:::

```{r}
# Figure 6C-1
f_caption_figure("Highest level of education for people who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by LGA, 2021")

education_qualification_lga %>%
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  #
  f_graph_grouped(x_variable = lga_name,
                  #
                  factor_levels  = c("postgrad_n", "undergrad_n", "certificate_n", "year_10_12_n", "year_9_n"),
                  factor_colours = c(colour_lightblue, colour_green, colour_orange, colour_yellow, colour_burgundy),
                  factor_labels  = c("Postgraduate", "Undergraduate", "Certificate and Diploma", "Year 10-12", "Year 9 and below"),
                  #
                  y_breaks = 200,
                  y_expand = 100) +
  #
  guides(fill = guide_legend(reverse = FALSE))
```

#### Data notes:
 
::: {custom-style="Footnotes"}
- Data are obtained from the [ABS Census 2021](https://www.abs.gov.au/census) and are publicly available at Local Government Area (LGA), ABS Statistical Area Level 2 (SA2), and Indigenous Area of Residence (IARE) level.
 
- Aboriginal and/or Torres Strait Islander peoples are defined as those who identified themselves as Aboriginal and/or Torres Strait Islander in the ABS Census 2021.
 
- Reported counts and percentages are a likely underestimate due to under-identification.

- The population for this indicator is people aged 15 years and over who are no longer attending primary or secondary school.

- People who did not attend school are included in the 'Year 9 and below' category.
:::

<!---BLOCK_LANDSCAPE_START--->

```{r}
# Table 6C-1
f_caption_table("Highest level of education for people who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by LGA, 2021")

education_qualification_lga %>% 
  dplyr::filter(lga_name %in% c(nephu_lga, "NEPHU")) %>%
  #
  dplyr::mutate(postgrad_prop    = round(postgrad_prop, digits = 1),
                undergrad_prop   = round(undergrad_prop, digits = 1),
                certificate_prop = round(certificate_prop, digits = 1),
                year_10_12_prop  = round(year_10_12_prop, digits = 1),
                year_9_prop      = round(year_9_prop, digits = 1)) %>% 
  #
  f_table_grouped_lga(table_headers = c("n", "%", "n", "%", "n", "%", "n", "%", "n", "%"),
                      group_headers = c("Postgraduate", "Undergraduate", "Certificate and Diploma", 
                                        "Year 10-12", "Year 9 and below"),
                      group_widths  = c(2, 2, 2, 2, 2),
                      last_column   = 12)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
- Data are obtained from the [ABS Census 2021](https://www.abs.gov.au/census) and are publicly available at Local Government Area (LGA), ABS Statistical Area Level 2 (SA2), and Indigenous Area of Residence (IARE) level.
 
- Aboriginal and/or Torres Strait Islander peoples are defined as those who identified themselves as Aboriginal and/or Torres Strait Islander in the ABS Census 2021.
 
- Reported counts and percentages are a likely underestimate due to under-identification.

- The population for this indicator is people aged 15 years and over who are no longer attending primary or secondary school.

- People who did not attend school are included in the 'Year 9 and below' category.
:::

<!---BLOCK_LANDSCAPE_STOP--->

```{r}
# Figure 6C-2
f_caption_figure("Highest level of education for NEPHU residents who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by age and sex, 2021")

education_qualification_agesex %>%
  dplyr::filter(age_group %in% age_group_lvl) %>% 
  #
  f_graph_grouped(x_variable = age_group,
                  #
                  factor_levels  = c("postgrad_n", "undergrad_n", "certificate_n", "year_10_12_n", "year_9_n"),
                  factor_colours = c(colour_lightblue, colour_green, colour_orange, colour_yellow, colour_burgundy),
                  factor_labels  = c("Postgraduate", "Undergraduate", "Certificate and Diploma", 
                                     "Year 10-12", "Year 9 and below"),
                  #
                  y_breaks = 100,
                  y_expand = 75) +
  #
  facet_wrap(~sex,
             axes = "all") +
  #
  guides(fill = guide_legend(reverse = FALSE))
```

#### Data notes:
 
::: {custom-style="Footnotes"}
- Data are obtained from the [ABS Census 2021](https://www.abs.gov.au/census) and are publicly available at Local Government Area (LGA), ABS Statistical Area Level 2 (SA2), and Indigenous Area of Residence (IARE) level.
 
- Aboriginal and/or Torres Strait Islander peoples are defined as those who identified themselves as Aboriginal and/or Torres Strait Islander in the ABS Census 2021.
 
- Reported counts and percentages are a likely underestimate due to under-identification.

- People with an unknown or not stated age or sex have been excluded from age/sex breakdowns.

- The population for this indicator is people aged 15 years and over who are no longer attending primary or secondary school.

- People who did not attend school are included in the 'Year 9 and below' category.
:::

<!---BLOCK_LANDSCAPE_START--->

```{r}
# Table 6C-2
f_caption_table("Highest level of education for females who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by age, 2021")

education_qualification_agesex %>% 
  dplyr::mutate(postgrad_prop    = round(postgrad_prop, digits = 1),
                undergrad_prop   = round(undergrad_prop, digits = 1),
                certificate_prop = round(certificate_prop, digits = 1),
                year_10_12_prop  = round(year_10_12_prop, digits = 1),
                year_9_prop      = round(year_9_prop, digits = 1)) %>% 
  #
  f_table_grouped_agesex(sex_name = "Female",
                         #
                         table_headers = c("n", "%", "n", "%", "n", "%", "n", "%", "n", "%"),
                         group_headers = c("Postgraduate", "Undergraduate", "Certificate and Diploma", 
                                           "Year 10-12", "Year 9 and below"),
                         group_widths  = c(2, 2, 2, 2, 2),
                         last_column   = 12)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
- Data are obtained from the [ABS Census 2021](https://www.abs.gov.au/census) and are publicly available at Local Government Area (LGA), ABS Statistical Area Level 2 (SA2), and Indigenous Area of Residence (IARE) level.
 
- Aboriginal and/or Torres Strait Islander peoples are defined as those who identified themselves as Aboriginal and/or Torres Strait Islander in the ABS Census 2021.
 
- Reported counts and percentages are a likely underestimate due to under-identification.

- People with an unknown or not stated age or sex have been excluded from age/sex breakdowns.

- The population for this indicator is people aged 15 years and over who are no longer attending primary or secondary school.

- People who did not attend school are included in the 'Year 9 and below' category.
:::

\newpage

```{r}
# Table 6C-3
f_caption_table("Highest level of education for males who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by age, 2021")

education_qualification_agesex %>% 
  dplyr::mutate(postgrad_prop    = round(postgrad_prop, digits = 1),
                undergrad_prop   = round(undergrad_prop, digits = 1),
                certificate_prop = round(certificate_prop, digits = 1),
                year_10_12_prop  = round(year_10_12_prop, digits = 1),
                year_9_prop      = round(year_9_prop, digits = 1)) %>% 
  #
  f_table_grouped_agesex(sex_name = "Male",
                         #
                         table_headers = c("n", "%", "n", "%", "n", "%", "n", "%", "n", "%"),
                         group_headers = c("Postgraduate", "Undergraduate", "Certificate and Diploma", 
                                           "Year 10-12", "Year 9 and below"),
                         group_widths  = c(2, 2, 2, 2, 2),
                         last_column   = 12)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
- Data are obtained from the [ABS Census 2021](https://www.abs.gov.au/census) and are publicly available at Local Government Area (LGA), ABS Statistical Area Level 2 (SA2), and Indigenous Area of Residence (IARE) level.
 
- Aboriginal and/or Torres Strait Islander peoples are defined as those who identified themselves as Aboriginal and/or Torres Strait Islander in the ABS Census 2021.
 
- Reported counts and percentages are a likely underestimate due to under-identification.

- People with an unknown or not stated age or sex have been excluded from age/sex breakdowns.

- The population for this indicator is people aged 15 years and over who are no longer attending primary or secondary school.

- People who did not attend school are included in the 'Year 9 and below' category.
:::

<!---BLOCK_LANDSCAPE_STOP--->

# Tertiary qualifications

::: {custom-style="Dotpt"}
This indicator aligns with Closing the Gap Target 6: By 2031, increase the proportion of Aboriginal and Torres Strait Islander people (aged 25-34 years) who have completed a tertiary qualification (Certificate III and above) to 70%.
:::

<br>

```{r, include = FALSE}
# ------------------------------------------------------------------------------
# Tertiary qualifications - trend over time ----
# Data from ABS Census via CtG Information Repository
# ------------------------------------------------------------------------------
# Read in ABS data
education_uni_trend <- readxl::read_excel(file.path(root_folder, subfolder_ctg,
                                                    "CTG06_Education_Pathways_Data_Tables.xlsx"),
                                          skip  = 1,
                                          sheet = "Table CtG6A.1") %>% 
  janitor::clean_names() %>% 
  #
  dplyr::rename(indigenous_status = x3,
                year = x4) %>%
  #
  tidyr::fill(indigenous_status, year) %>%
  #
  dplyr::filter(!is.na(vic)) %>% 
  dplyr::filter(row_number() %in% 1:10) %>%
  #
  dplyr::select(indigenous_status,
                year,
                vic)

# Data wrangling
education_uni_trend <- education_uni_trend %>% 
  dplyr::mutate(vic = as.numeric(vic))

# ------------------------------------------------------------------------------
# Tertiary qualifications - young adults
# Data from ABS Census using TableBuilderPro
# ------------------------------------------------------------------------------
# Data wrangling
education_uni_youth <- f_read_census_agesex("Aboriginal_Education_Qualification_YoungAdult.csv",
                                            variable = "qualification") %>% 
  #
  dplyr::filter(!qualification %in% c("Not stated", "Not applicable", "Supplementary Codes")) %>%
  #
  dplyr::mutate(qualification = dplyr::case_when(
    qualification == "Postgraduate Degree Level"                       ~ "yes",
    qualification == "Graduate Diploma and Graduate Certificate Level" ~ "yes",
    qualification == "Bachelor Degree Level"                           ~ "yes",
    qualification == "Advanced Diploma and Diploma Level"              ~ "yes",
    qualification == "Certificate III & IV Level"                      ~ "yes",
    #
    TRUE ~ "no"))

# Data wrangling - by LGA
education_uni_youth_lga <- education_uni_youth %>% 
  f_wrangle_census_lga(variable = qualification,
                       level    = "person",
                       dichot   = "yes")

# Data wrangling - by sex
education_uni_youth_sex <- education_uni_youth %>%
  f_wrangle_census_agesex(variable = qualification,
                          dichot   = "yes") %>% 
  #
  dplyr::filter(age_group == "Total") %>% 
  #
  dplyr::bind_rows(education_uni_youth_lga %>% 
                     dplyr::filter(lga_name == "NEPHU") %>%
                     #
                     dplyr::mutate(age_group = "NEPHU",
                                   sex       = "NEPHU") %>% 
                     #
                     dplyr::select(-lga_name))
```


```{r, include = FALSE}
#
results <- list()
data    <- education_uni_youth_lga

results$nephu_n <- data %>% 
  dplyr::filter(lga_name == "NEPHU") %>% 
  pull(aboriginal_n)

results$nephu_prop <- data %>% 
  dplyr::filter(lga_name == "NEPHU") %>% 
  pull(aboriginal_prop)

results$non_prop <- data %>% 
  dplyr::filter(lga_name == "NEPHU") %>% 
  pull(non_aboriginal_prop)

results$comp_non <- dplyr::case_when(
  results$nephu_prop < results$non_prop * 0.9 ~ "lower than for",
  results$nephu_prop > results$non_prop * 1.1 ~ "higher than for",
  TRUE ~ "similar to")

results$rank_n <- data %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  dplyr::arrange(desc(aboriginal_n)) %>% 
  dplyr::select(lga_name,
                aboriginal_n)

results$rank_prop <- data %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  dplyr::arrange(desc(aboriginal_prop)) %>% 
  dplyr::select(lga_name,
                aboriginal_prop)

results$rank_diff <- data %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  dplyr::arrange(desc(aboriginal_difference)) %>% 
  dplyr::select(lga_name,
                aboriginal_difference)
```

::: {custom-style="Dotpt"}
- The percentage of Aboriginal and/or Torres Strait Islander people aged 25-34 years living in Victoria with a tertiary qualification (Certificate III and above) increased from 29% in 2001 to 59% in 2021.

- Overall, `r round(results$nephu_prop, digits = 0)`% (n = `r formatC(results$nephu_n, big.mark = ",")`) of people living in the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census have a tertiary qualification. This is  `r results$comp_non` non-Aboriginal and/or Torres Strait Islander people living in the NEPHU catchment (`r round(results$non_prop, digits = 0)`%).

- The number of people living in the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census who have a tertiary qualification ranged from 186 (Darebin LGA) to 15 (Manningham LGA).

- The gap between Aboriginal and/or Torres Strait Islander and non-Aboriginal and/or Torres Strait Islander people was biggest for Darebin LGA (57 fewer people) and Hume LGA (43 fewer people).

- The percentage of people living in the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census who have a tertiary qualification ranged from 91% (Nillumbik LGA) to 56% (Knox LGA).

- 4/12 LGAs in the NEPHU catchment were above the Closing the Gap target of 70%.

- 71% of females living in the NEPHU catchment aged 25-34 years who identified as Aboriginal and/or Torres Strait Islander in the Census have a tertiary qualification, compared to 59% of males.
:::

\newpage

```{r}
# Figure 6C-3
f_caption_figure("Percentage of people aged 25-34 years who identified as Aboriginal and/or Torres Strait Islander and who have completed a tertiary qualification (Certificate III and above), Victoria, by Aboriginal and/or Torres Strait Islander status, 2001-2021")

education_uni_trend %>% 
  f_graph_line_comparison(y_variable = vic,
                          y_max      = 100,
                          y_breaks   = 10,
                          y_expand   = 1,
                          y_title    = "Percentage of people") %>% 
  #
  f_graph_ctg_target(ctg_target = 70)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
- Data are obtained from the [Closing the Gap Information Repository](https://www.pc.gov.au/closing-the-gap-data), compiled from the [ABS Census](https://www.abs.gov.au/census) for the Census years between 2001 and 2021.
 
- Aboriginal and/or Torres Strait Islander peoples are defined as those who identified themselves as Aboriginal and/or Torres Strait Islander in each individual ABS Census.
 
- Reported counts and percentages are a likely underestimate due to under-identification.

- The Closing the Gap Target (70%) is marked with a dashed blue line.
:::

```{r}
# Table 6C-4
f_caption_table("Percentage of people aged 25-34 years who identified as Aboriginal and/or Torres Strait Islander and who have completed a tertiary qualification (Certificate III and above), Victoria, by Aboriginal and/or Torres Strait Islander status, 2001-2021")

education_uni_trend %>% 
  tidyr::pivot_wider(names_from  = "indigenous_status",
                     values_from = "vic") %>% 
  #
  dplyr::arrange(year) %>% 
  #
  f_table_rate_year() %>% 
  #
  flextable::set_header_labels(values = c("Year", "Yes (%)", "No (%)"))
```

#### Data notes:
 
::: {custom-style="Footnotes"}
- Data are obtained from the [Closing the Gap Information Repository](https://www.pc.gov.au/closing-the-gap-data), compiled from the [ABS Census](https://www.abs.gov.au/census) for the Census years between 2001 and 2021.
 
- Aboriginal and/or Torres Strait Islander peoples are defined as those who identified themselves as Aboriginal and/or Torres Strait Islander in each individual ABS Census.
 
- Reported counts and percentages are a likely underestimate due to under-identification.
:::

\newpage

```{r}
# Figure 6C-4
f_caption_figure("(A) Number and (B) percentage of people aged 25-34 years who identified as Aboriginal and/or Torres Strait Islander and who have completed a tertiary qualification (Certificate III and above), NEPHU catchment, by LGA, 2021")

figure_number <- education_uni_youth_lga %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  #
  f_graph_lga(y_variable = aboriginal_n,
              y_max      = NA,
              y_breaks   = 20,
              y_expand   = 15,
              y_title    = "Number of people",
              plot_tag   = "A")

figure_percent <- education_uni_youth_lga %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  #
  f_graph_lga(y_variable = aboriginal_prop,
              y_max      = 100,
              y_breaks   = 10,
              y_expand   = 1,
              y_title    = "Percentage of people",
              plot_tag   = "B") %>% 
  #
  f_graph_ctg_target(ctg_target = 70)

figure_number + figure_percent
```

#### Data notes:
 
::: {custom-style="Footnotes"}
- Data are obtained from the [ABS Census 2021](https://www.abs.gov.au/census) and are publicly available at Local Government Area (LGA), ABS Statistical Area Level 2 (SA2), and Indigenous Area of Residence (IARE) level.
 
- Aboriginal and/or Torres Strait Islander peoples are defined as those who identified themselves as Aboriginal and/or Torres Strait Islander in the ABS Census 2021.
 
- Reported counts and percentages are a likely underestimate due to under-identification.

- The Closing the Gap Target (70%) is marked with a dashed blue line in Chart B.
:::

```{r}
# Figure 6C-5
f_caption_figure("Observed vs. expected number of people aged 25-34 years who have completed a tertiary qualification (Certificate III and above) and who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by LGA, 2021")

education_uni_youth_lga %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  #
  f_graph_expected(variable = lga_name,
                   y_breaks = 25,
                   y_expand = 10)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
- Data are obtained from the [ABS Census 2021](https://www.abs.gov.au/census) and are publicly available at Local Government Area (LGA), ABS Statistical Area Level 2 (SA2), and Indigenous Area of Residence (IARE) level.
 
- Aboriginal and/or Torres Strait Islander peoples are defined as those who identified themselves as Aboriginal and/or Torres Strait Islander in the ABS Census 2021.
 
- Reported counts and percentages are a likely underestimate due to under-identification.
:::

```{r}
# Table 6C-5
f_caption_table("Observed vs. expected number of people aged 25-34 years who have completed a tertiary qualification (Certificate III and above) and who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by LGA, 2021")

education_uni_youth_lga %>% 
  dplyr::filter(lga_name %in% c(nephu_lga, "NEPHU")) %>%
  #
  f_table_expected(variable       = lga_name,
                   variable_label = "LGA",
                   group_label    = " ",
                   total_row      = 13)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
- Data are obtained from the [ABS Census 2021](https://www.abs.gov.au/census) and are publicly available at Local Government Area (LGA), ABS Statistical Area Level 2 (SA2), and Indigenous Area of Residence (IARE) level.
 
- Aboriginal and/or Torres Strait Islander peoples are defined as those who identified themselves as Aboriginal and/or Torres Strait Islander in the ABS Census 2021.
 
- Reported counts and percentages are a likely underestimate due to under-identification.
:::

```{r}
# Table 6C-6
f_caption_table("Observed vs. expected number of people aged 25-34 years who have completed a tertiary qualification (Certificate III and above) and who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by sex, 2021")

education_uni_youth_sex %>%
  f_table_expected(variable       = sex,
                   variable_label = "Sex",
                   group_label    = " ",
                   total_row      = 3)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
- Data are obtained from the [ABS Census 2021](https://www.abs.gov.au/census) and are publicly available at Local Government Area (LGA), ABS Statistical Area Level 2 (SA2), and Indigenous Area of Residence (IARE) level.
 
- Aboriginal and/or Torres Strait Islander peoples are defined as those who identified themselves as Aboriginal and/or Torres Strait Islander in the ABS Census 2021.
 
- Reported counts and percentages are a likely underestimate due to under-identification.

- People with an unknown or not stated age or sex have been excluded from age/sex breakdowns.
:::
