==> Code/08d_Other_Discrimination.Rmd <==
---
title: "NEPHU First Nations Population Profile"
subtitle: "Other Determinants of Health: Discrimination"
output:
  officedown::rdocx_document: 
    page_size:
      width: 8.27
      length: 11.69
    page_margins:
      bottom: 0.5
      top: 0.5
      right: 0.8
      left: 0.8
      gutter: 0
    reference_docx: Template.docx
    
knit: (function(input, ...) {
  rmarkdown::render(input, output_file = paste0("NEPHU First Nations Profile - ", "08D Other Discrimination ", format(Sys.Date(), "%Y-%m-%d")), output_dir = "../Output/Dev") })
---

```{r setup, include = FALSE}
#
knitr::opts_chunk$set(echo    = FALSE, 
                      warning = FALSE, 
                      message = FALSE,
                      #
                      fig.align  = "center",
                      fig.width  = 7,
                      fig.height = 3.5,
                      #
                      ft.align = "center",
                      #
                      dpi = 300,
                      #
                      fig.topcaption = TRUE,
                      #
                      cache = FALSE)

if (!require("pacman")) install.packages("pacman")

pacman::p_load(here,
               tidyverse,
               rlang,
               janitor,
               readxl,
               sf,
               flextable,
               colorRamps,
               officer,
               officedown,
               extrafont,
               ggnewscale,
               patchwork)

# Uncomment and run these lines if you get any weird "missing font" errors
#extrafont::font_import()
#extrafont::loadfonts(device = "win")
```


```{r, include = FALSE}
#
source(here::here("Code", "00_Setup.R"))

source(here::here("Code", "Functions", "Data prep", "f_read_natsihs_sex.R"))

source(here::here("Code", "Functions", "Graphs", "f_graph_line_comparison.R"))
source(here::here("Code", "Functions", "Graphs", "f_graph_survey_sex.R"))
 
source(here::here("Code", "Functions", "Tables", "f_table_survey_sex.R"))
```

**Table of Contents**

```{r}
#
block_toc(level = 2)
```

\newpage

```{r, include = FALSE}
# Numbering and styling for figures and tables
f_caption_figure <- function(title_text) {
  
  caption <- block_caption(title_text, 
                           style   = "caption", 
                           autonum = run_autonum(seq_id    = "Figure", 
                                                 pre_label = "Figure 8D-"))
  
  return(caption)
  
}

f_caption_table <- function(title_text) {
  
  caption <- block_caption(title_text, 
                           style   = "caption", 
                           autonum = run_autonum(seq_id    = "Table", 
                                                 pre_label = "Table 8D-"))
  
  return(caption)
  
}
```

### What external factors may influence how Aboriginal and/or Torres Strait Islander people and families in the NEPHU catchment are supported to prioritise health and wellbeing, particularly in connection to caring for children and keeping families strong?

# Racial Prejudice

```{r}
# ------------------------------------------------------------------------------
# Discrimination: experience of racial prejudice
# Data from Australian Reconciliation Barometer
# ------------------------------------------------------------------------------
# Read in Australian Reconciliation Barometer data
discrimination_prejudice <- readxl::read_excel(file.path(root_folder, subfolder_ctg, "CTG14_Wellbeing_Data_Tables.xlsx"),
                                               skip  = 1,
                                               sheet = "Table SE14g.1") %>%
  janitor::clean_names() %>%
  #
  dplyr::rename(year = x4,
                type = x2,
                indigenous_status = x3) %>%
  #
  tidyr::fill(type, indigenous_status) %>% 
  #
  dplyr::filter(!is.na(vic) & type != "Total population aged 18 years or over (weighted estimate)")

# Data wrangling
discrimination_prejudice <- discrimination_prejudice %>%
  dplyr::select(year,
                indigenous_status,
                type,
                value = vic) %>%
  #
  dplyr::mutate(value = as.numeric(value),
                #
                indigenous_status = dplyr::case_when(
                  indigenous_status == "Aboriginal and Torres Strait Islander people" ~ "aboriginal",
                  indigenous_status == "General community" ~ "non_aboriginal",
                  TRUE ~ NA_character_),
                #
                type = dplyr::case_when(
                  type == "Proportion (b)" ~ "prop",
                  type == "Number (weighted estimate)" ~ "n",
                  TRUE ~ NA_character_))
```

::: {custom-style="Dotpt"}
- Data for this indicator were obtained from the Australian Reconciliation Barometer and are only available at a statewide level.

- In 2022, 60% of Aboriginal and/or Torres Strait Islander survey respondents living in Victoria reported that they had experienced racial prejudice during the previous six months, increasing from 51% in 2018 but lower than for 2020 (71%).

- The percentage of people who had experienced racial prejudice was higher for Aboriginal and/or Torres Strait Islander people than for non-Indigenous people at each time point.
:::

\newpage

```{r}
# Figure 8D-1
f_caption_figure("Percentage of people reporting that they experienced racial prejudice in the previous six months, Victoria, by Aboriginal and/or Torres Strait Islander status and year, 2018-2022")

discrimination_prejudice %>% 
  dplyr::filter(type == "prop") %>% 
  #
  f_graph_line_comparison(y_variable = value,
                          y_max      = 100,
                          y_breaks   = 10,
                          y_expand   = 1,
                          y_title    = "Percentage of people")
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_reconciliation)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_survey_estimate)`

`r knitr::asis_output(note_caveat_unknown_survey)`

- 'Racial prejudice' is defined as a survey respondent experiencing any of the following on the basis of their race:

    - Verbal abuse
    
    - Social media abuse
    
    - Physical violence
    
    - Prevented from renting or buying a property
    
    - Refused service in a shop
    
    - Refused entry to a venue
    
    - Other prejudice or discrimination
:::

\newpage

```{r}
# Table 8D-1
f_caption_table("Experience of racial prejudice in the previous six months, Victoria, by Aboriginal and/or Torres Strait Islander status and year, 2018-2022")

discrimination_prejudice %>% 
  tidyr::pivot_wider(names_from  = type,
                     values_from = value) %>% 
  #
  dplyr::mutate(prop = sprintf("%.1f", prop)) %>% 
  #
  tidyr::pivot_wider(names_from  = indigenous_status,
                     values_from = c(n, prop)) %>% 
  #
  dplyr::select(year,
                n_aboriginal,
                prop_aboriginal,
                n_non_aboriginal,
                prop_non_aboriginal) %>% 
  #
  dplyr::arrange(year) %>% 
  #
  flextable::flextable() %>%
  #
  flextable::set_header_labels(values = c("Year", "Yes\n(n)", "Yes\n(%)", "No\n(n)", "No\n(%)")) %>%
  #
  flextable::add_header_row(values    = c(" ", "Identified as Aboriginal and/or Torres Strait Islander"),
                            colwidths = c(1, 4)) %>%
  #
  flextable::valign(valign = "bottom",
                    part   = "header") %>%
  #
  flextable::align(j = 1,
                   align = "left",
                   part  = "all") %>%
  #
  flextable::align(j = c(2:5),
                   align = "right",
                   part  = "all") %>%
  #
  flextable::fontsize(size = 10,
                      part = "all") %>%
  #
  flextable::bold(part = "header") %>%
  #
  flextable::width(j = 1:5,
                   width = 1) %>%
  #
  flextable::line_spacing(space = 0.75,
                          part  = "body")
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_reconciliation)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_survey_estimate)`

`r knitr::asis_output(note_caveat_unknown_survey)`

- 'Racial prejudice' is defined as a survey respondent experiencing any of the following on the basis of their race:

    - Verbal abuse
    
    - Social media abuse
    
    - Physical violence
    
    - Prevented from renting or buying a property
    
    - Refused service in a shop
    
    - Refused entry to a venue
    
    - Other prejudice or discrimination
:::

\newpage

# Unfair Treatment

```{r}
# ------------------------------------------------------------------------------
# Discrimination: experience of unfair treatment
# Data from ABS National Aboriginal and Torres Strait Islander Health Survey
# ------------------------------------------------------------------------------
# Read in NATSIHS data
discimination_unfair <- f_read_natsihs_sex("Indicator_Vic_Sex_Discrimination_Unfair.csv")

# Data wrangling
discimination_unfair <- discimination_unfair %>% 
  dplyr::rowwise() %>% 
  dplyr::mutate(answered_n = sum(across(had_unfair_treatment_n:has_not_had_unfair_treatment_n))) %>% 
  dplyr::ungroup() %>% 
  #
  janitor::adorn_totals(fill = "Total") %>% 
  #
  dplyr::rename(unfair_n    = had_unfair_treatment_n,
                notunfair_n = has_not_had_unfair_treatment_n) %>% 
  #
  dplyr::mutate(across(unfair_n:notunfair_n,
                       ~ . / answered_n * 100,
                       .names = "{.col}_prop"),
                #
                answered_prop = answered_n / answered_n * 100) %>% 
  #
  dplyr::rename_with(~ stringr::str_replace(., "_n_prop$", "_prop"), ends_with("_n_prop")) %>% 
  #
  dplyr::select(sex,
                starts_with("unfair"),
                starts_with("notunfair"),
                starts_with("answered"))

# ------------------------------------------------------------------------------
# Discrimination: avoidance of situations due to previous unfair treatment
# Data from ABS National Aboriginal and Torres Strait Islander Health Survey
# ------------------------------------------------------------------------------
# Read in NATSIHS data
discimination_avoid <- f_read_natsihs_sex("Indicator_Vic_Sex_Discrimination_Avoided.csv")

# Data wrangling
discimination_avoid <- discimination_avoid %>% 
  dplyr::rowwise() %>% 
  dplyr::mutate(answered_n = sum(across(avoided_situations_due_to_past_unfair_treatment_in_last_12_months_n:did_not_avoid_situations_due_to_past_unfair_treatment_in_last_12_months_n))) %>% 
  dplyr::ungroup() %>% 
  #
  janitor::adorn_totals(fill = "Total") %>% 
  #
  dplyr::rename(avoid_n    = avoided_situations_due_to_past_unfair_treatment_in_last_12_months_n,
                notavoid_n = did_not_avoid_situations_due_to_past_unfair_treatment_in_last_12_months_n) %>% 
  #
  dplyr::mutate(across(avoid_n:notavoid_n,
                       ~ . / answered_n * 100,
                       .names = "{.col}_prop"),
                #
                answered_prop = answered_n / answered_n * 100) %>% 
  #
  dplyr::rename_with(~ stringr::str_replace(., "_n_prop$", "_prop"), ends_with("_n_prop")) %>% 
  #
  dplyr::select(sex,
                starts_with("avoid"),
                starts_with("notavoid"),
                starts_with("answered"))
```

::: {custom-style="Dotpt"}
- Data for this indicator were obtained from the 2022-23 ABS National Aboriginal and Torres Strait Islander Health Survey.

- Overall, 25% of Aboriginal and/or Torres Strait Islander survey respondents aged 15 years and over living in Greater Melbourne reported that they had experienced unfair treatment due to their race in the 12 months prior to the survey. This equates to approximately 6,200 Aboriginal and/or Torres Strait Islander people in Greater Melbourne.

- Female survey respondents were more likely to have experienced unfair treatment than males (28% vs. 23%).

- 12% of Aboriginal and/or Torres Strait Islander survey respondents aged 15 years and over living in Greater Melbourne reported that they had avoided situations due to previous experiences of unfair treatment. This equates to approximately 3,000 Aboriginal and/or Torres Strait Islander people in Greater Melbourne.

- Males were more likely to have avoided situations due to previous experiences of unfair treatment than females (16% vs. 8.0%).
:::

```{r}
# Table 8D-2
f_caption_table("Experience of unfair treatment in the 12 months prior to the survey for survey respondents who identified as Aboriginal and/or Torres Strait Islander, Greater Melbourne, by sex, 2022-2023")

x_labels <- c("unfair"    = "Yes",
              "notunfair" = "No",
              "answered"  = "Total")

discimination_unfair %>% 
  f_table_survey_sex(x_labels       = x_labels,
                     variable_label = "Unfair treatment",
                     total_row      = 3) %>% 
  #
  flextable::colformat_double(j = c(3, 5, 7),
                              digits = 1)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_natsihs)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_survey_estimate)`

`r knitr::asis_output(note_caveat_unknown_survey)`
:::

\newpage

```{r}
# Table 8D-3
f_caption_table("Avoidance of situations due to previous unfair treatment for survey respondents who identified as Aboriginal and/or Torres Strait Islander, Greater Melbourne, by sex, 2022-2023")

x_labels <- c("avoid"    = "Yes",
              "notavoid" = "No",
              "answered" = "Total")

discimination_avoid %>% 
  f_table_survey_sex(x_labels       = x_labels,
                     variable_label = "Avoided situation",
                     total_row      = 3) %>% 
  #
  flextable::colformat_double(j = c(3, 5, 7),
                              digits = 1)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_natsihs)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_survey_estimate)`

`r knitr::asis_output(note_caveat_unknown_survey)`
:::
