==> Code/10g_Consequences_ChronicHeart.Rmd <==
---
title: "NEPHU First Nations Population Profile"
subtitle: "Health Consequences: Heart and Vascular Disease"
output:
  officedown::rdocx_document: 
    page_size:
      width: 8.27
      length: 11.69
    page_margins:
      bottom: 0.5
      top: 0.5
      right: 0.8
      left: 0.8
      gutter: 0
    reference_docx: Template.docx
    
knit: (function(input, ...) {
  rmarkdown::render(input, output_file = paste0("NEPHU First Nations Profile - ", "10G Consequences Heart Disease ", format(Sys.Date(), "%Y-%m-%d")), output_dir = "../Output/Dev") })
---

```{r setup, include = FALSE}
#
knitr::opts_chunk$set(echo    = FALSE, 
                      warning = FALSE, 
                      message = FALSE,
                      #
                      fig.align  = "center",
                      fig.width  = 7,
                      fig.height = 3.5,
                      #
                      ft.align = "center",
                      #
                      dpi = 300,
                      #
                      fig.topcaption = TRUE,
                      #
                      cache = FALSE)

if (!require("pacman")) install.packages("pacman")

pacman::p_load(here,
               tidyverse,
               rlang,
               janitor,
               readxl,
               sf,
               flextable,
               colorRamps,
               officer,
               officedown,
               extrafont,
               ggnewscale,
               patchwork)

# Uncomment and run these lines if you get any weird "missing font" errors
#extrafont::font_import()
#extrafont::loadfonts(device = "win")
```


```{r, include = FALSE}
#
source(here::here("Code", "00_Setup.R"))

source(here::here("Code", "Functions", "Data prep", "f_read_census_lga.R"))
source(here::here("Code", "Functions", "Data prep", "f_wrangle_census_lga.R"))
source(here::here("Code", "Functions", "Data prep", "f_read_census_agesex.R"))
source(here::here("Code", "Functions", "Data prep", "f_wrangle_census_agesex.R"))
source(here::here("Code", "Functions", "Data prep", "f_read_phidu_iare_hospital.R"))

source(here::here("Code", "Functions", "Graphs", "f_graph_lga.R"))
source(here::here("Code", "Functions", "Graphs", "f_graph_agesex.R"))
source(here::here("Code", "Functions", "Graphs", "f_graph_expected.R"))
source(here::here("Code", "Functions", "Graphs", "f_graph_iare.R"))

source(here::here("Code", "Functions", "Tables", "f_table_expected.R"))
source(here::here("Code", "Functions", "Tables", "f_table_iare.R"))
```

**Table of Contents**

```{r}
#
block_toc(level = 2)
```

\newpage

```{r, include = FALSE}
# Numbering and styling for figures and tables
f_caption_figure <- function(title_text) {
  
  caption <- block_caption(title_text, 
                           style   = "caption", 
                           autonum = run_autonum(seq_id    = "Figure", 
                                                 pre_label = "Figure 10G-"))
  
  return(caption)
  
}

f_caption_table <- function(title_text) {
  
  caption <- block_caption(title_text, 
                           style   = "caption", 
                           autonum = run_autonum(seq_id    = "Table", 
                                                 pre_label = "Table 10G-"))
  
  return(caption)
  
}
```

### What chronic conditions affect Aboriginal and/or Torres Strait Islander people living in the NEPHU catchment?

# Heart Disease

## Prevalence

```{r, include = FALSE}
# ------------------------------------------------------------------------------
# Heart disease, by LGA
# Data from ABS Census using TableBuilderPro
# ------------------------------------------------------------------------------
# Read in Census data
health_heart_lga <- f_read_census_lga("Aboriginal_Health_Heart_LGA.csv",
                                      variable = "heart") %>% 
  #
  dplyr::filter(heart != "Not stated")

# Data wrangling
health_heart_lga <- health_heart_lga %>%
  #
  dplyr::mutate(heart = dplyr::case_when(
    heart == "Has heart disease (including heart attack or angina)" ~ "yes",
    TRUE ~ "no")) %>% 
  #
  f_wrangle_census_lga(variable = heart,
                       level    = "person",
                       dichot   = "yes")

# ------------------------------------------------------------------------------
# Heart disease, by age and sex
# Data from ABS Census using TableBuilderPro
# ------------------------------------------------------------------------------
# Read in Census data
health_heart_agesex <- f_read_census_agesex("Aboriginal_Health_Heart_AgeSex.csv",
                                            variable = "heart") %>% 
  #
  dplyr::filter(heart != "Not stated")

# Data wrangling
health_heart_agesex <- health_heart_agesex %>% 
  #
  dplyr::mutate(heart = dplyr::case_when(
    heart == "Has heart disease (including heart attack or angina)" ~ "yes",
    TRUE ~ "no")) %>% 
  #
  f_wrangle_census_agesex(variable = heart,
                          dichot   = "yes")
```


```{r}
#
results <- list()
data    <- health_heart_lga

results$rank_n <- data %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  dplyr::arrange(desc(aboriginal_n)) %>% 
  dplyr::select(lga_name,
                aboriginal_n)

results$rank_prop <- data %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  dplyr::arrange(desc(aboriginal_prop)) %>% 
  dplyr::select(lga_name,
                aboriginal_prop)

results$rank_diff <- data %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  dplyr::arrange(desc(aboriginal_difference)) %>% 
  dplyr::select(lga_name,
                aboriginal_difference)
```

::: {custom-style="Dotpt"}
- Data for this indicator were obtained from the ABS Census 2021.

- Overall, 3.7% (n = 426) of people living in the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census have been told by a doctor or nurse that they have heart disease. This is similar to non-Indigenous people living in the NEPHU catchment (3.7%).

- The number of people living in different LGAs within the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census who have been told by a doctor or nurse that they have heart disease ranged from 68 (Whittlesea LGA) to 4 (Nillumbik LGA).

- The gap between Aboriginal and/or Torres Strait Islander and non-Indigenous people was biggest for Darebin and Knox LGAs.

- The percentage of people living in different LGAs within the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census who have been told by a doctor or nurse that they have heart disease ranged from 5.2% (Banyule LGA) to 1.1% (Nillumbik LGA).

- The number of people who have been told by a doctor or nurse that they have heart disease was higher for males than females, and the number and percentage of people who have been told by a doctor or nurse that they have heart disease increased with increasing age for both genders.
:::

\newpage

```{r}
# Figure 10G-1
f_caption_figure("(A) Number and (B) percentage of people who identified as Aboriginal and/or Torres Strait Islander and who have been told by a doctor or nurse that they have heart disease, NEPHU catchment, by LGA, 2021")

figure_number <- health_heart_lga %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  #
  f_graph_lga(y_variable = aboriginal_n,
              y_max      = NA,
              y_breaks   = 10,
              y_expand   = 3,
              y_title    = "Number of people",
              plot_tag   = "A")

figure_percent <- health_heart_lga %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  #
  f_graph_lga(y_variable = aboriginal_prop,
              y_max      = 10,
              y_breaks   = 1,
              y_expand   = 0.1,
              y_title    = "Percentage of people",
              plot_tag   = "B")

figure_number + figure_percent
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`

- 'Heart disease' includes heart attack and angina.
:::

\newpage

```{r}
# Figure 10G-2
f_caption_figure("Observed vs. expected number of Aboriginal and/or Torres Strait Islander people who have been told by a doctor or nurse that they have heart disease, NEPHU catchment, by LGA, 2021")

health_heart_lga %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  #
  f_graph_expected(variable = lga_name,
                   y_breaks = 10,
                   y_expand = 1)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`

- 'Heart disease' includes heart attack and angina.
:::

\newpage

```{r}
# Table 10G-1
f_caption_table("Observed vs. expected number of Aboriginal and/or Torres Strait Islander people who have been told by a doctor or nurse that they have heart disease, NEPHU catchment, by LGA, 2021")

health_heart_lga %>% 
  dplyr::filter(lga_name %in% c(nephu_lga, "NEPHU")) %>%
  #
  f_table_expected(variable       = lga_name,
                   variable_label = "LGA",
                   group_label    = " ",
                   total_row      = 13)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`

- 'Heart disease' includes heart attack and angina.
:::

\newpage

```{r}
# Figure 10G-3
f_caption_figure("(A) Number and (B) percentage of people who identified as Aboriginal and/or Torres Strait Islander and who have been told by a doctor or nurse that they have heart disease, NEPHU catchment, by age and sex, 2021")

figure_female_number <- health_heart_agesex %>%
  f_graph_agesex(sex        = "Female",
                 y_variable = aboriginal_n,
                 y_max      = 70,
                 y_breaks   = 10,
                 y_expand   = 1,
                 y_title    = "Number of females",
                 plot_title = "Females",
                 plot_tag   = "A")

figure_female_percent <- health_heart_agesex %>%
  f_graph_agesex(sex        = "Female",
                 y_variable = aboriginal_prop,
                 y_max      = 50,
                 y_breaks   = 5,
                 y_expand   = 0.5,
                 y_title    = "Percentage of females",
                 plot_title = " ",
                 plot_tag   = "B")

figure_male_number <- health_heart_agesex %>%
  f_graph_agesex(sex        = "Male",
                 y_variable = aboriginal_n,
                 y_max      = 70,
                 y_breaks   = 10,
                 y_expand   = 1,
                 y_title    = "Number of males",
                 plot_title = "Males",
                 plot_tag   = "A")

figure_male_percent <- health_heart_agesex %>%
  f_graph_agesex(sex        = "Male",
                 y_variable = aboriginal_prop,
                 y_max      = 50,
                 y_breaks   = 5,
                 y_expand   = 0.5,
                 y_title    = "Percentage of males",
                 plot_title = " ",
                 plot_tag   = "B")

figure_female_number + figure_female_percent
figure_male_number + figure_male_percent
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_unknown_agesex)`

- 'Heart disease' includes heart attack and angina.
:::

\newpage

```{r} 
# Figure 10G-4
f_caption_figure("Observed vs. expected number of Aboriginal and/or Torres Strait Islander people who have been told by a doctor or nurse that they have heart disease, NEPHU catchment, by age and sex, 2021") 

health_heart_agesex %>%  
  dplyr::filter(age_group %in% age_group_lvl) %>%  
  # 
  dplyr::mutate(age_group = factor(age_group, levels = age_group_lvl)) %>% 
  #
  f_graph_expected(variable = age_group, 
                   y_breaks = 10, 
                   y_expand = 7) +
  #
  facet_wrap(~sex,
             axes = "all")
``` 

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_unknown_agesex)`

- 'Heart disease' includes heart attack and angina.
:::

\newpage

```{r} 
# Table 10G-2
f_caption_table("Observed vs. expected number of Aboriginal and/or Torres Strait Islander people who have been told by a doctor or nurse that they have heart disease, NEPHU catchment, by age and sex, 2021") 

health_heart_agesex %>% 
  dplyr::filter(sex == "Female") %>%  
  # 
  f_table_expected(variable       = age_group,
                   variable_label = "Age group",
                   group_label    = "Females",
                   total_row      = 12)
  
health_heart_agesex %>% 
  dplyr::filter(sex == "Male") %>%  
  # 
  f_table_expected(variable       = age_group,
                   variable_label = "Age group",
                   group_label    = "Males",
                   total_row      = 12)
``` 

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_unknown_agesex)`

- 'Heart disease' includes heart attack and angina.
:::

\newpage

## Hospital Admissions - Ischaemic Heart Disease

```{r, include = FALSE}
# ------------------------------------------------------------------------------
# Hospitalisations - ischaemic heart disease
# Data from PHIDU Social Health Atlas
# ------------------------------------------------------------------------------
# Read in PHIDU data
hospital_ihd <- f_read_phidu_iare_hospital(data_sheet    = "Admissions_principal_diag",
                                           column_select = "admissions_for_ischaemic_heart_disease_aboriginal_people")

# Data wrangling
hospital_ihd <- hospital_ihd %>% 
  dplyr::mutate(iare_name = factor(iare_name, levels = iare_inc_vic_lvl),
                #
                hospital_n_label = round(hospital_n, digits = 0),
                #
                hospital_n_label = dplyr::case_when(
                  is.na(hospital_n) ~ "*",
                  TRUE ~ format(hospital_n_label, big.mark = ",")),
                #
                hospital_rate_label = round(hospital_rate, digits = 0),
                #
                hospital_rate_label = dplyr::case_when(
                  is.na(hospital_rate) ~ "*",
                  TRUE ~ format(hospital_rate_label, big.mark = ",")),
                #
                hospital_ratio_label = round(hospital_ratio, digits = 1),
                #
                hospital_ratio_label = dplyr::case_when(
                  is.na(hospital_ratio) ~ "*",
                  TRUE ~ format(hospital_ratio_label, big.mark = ","))) %>% 
  #
  dplyr::select(iare_name,
                iare_name_short,
                starts_with("hospital")) %>% 
  #
  dplyr::arrange(iare_name)
```


```{r}
#
results <- list()

data <- hospital_ihd %>% 
  dplyr::rename(n     = hospital_n,
                rate  = hospital_rate,
                ratio = hospital_ratio)

results$rank_n <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(n)) %>% 
  dplyr::select(iare_name,
                n)

results$rank_rate <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(rate)) %>% 
  dplyr::select(iare_name,
                rate)

results$rank_ratio <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(ratio)) %>% 
  dplyr::select(iare_name,
                ratio)
```

::: {custom-style="Dotpt"}
- Data for this indicator were obtained from the PHIDU Social Health Atlas, compiled from the Victorian Admitted Episodes Dataset.

- During the four-year period 01 July 2017 to 30 June 2021 there were 449 hospital admissions for ischaemic heart disease for Aboriginal and/or Torres Strait Islander people living in Greater Melbourne.

- The number of admissions for Aboriginal and/or Torres Strait Islander people living in different IAREs within the NEPHU catchment ranged from 76 (Northcote/Preston/Whittlesea) to 5 (Knox IARE).

- The age-standardised rate of admissions for Aboriginal and/or Torres Strait Islander people living in different IAREs within the NEPHU catchment ranged from 474 per 100,000 population (Northcote/Preston/Whittlesea IARE) to 100 per 100,000 population (Knox IARE).

- 4 of the 9 IAREs in the NEPHU catchment area with data available for reporting had an age-standardised admission rate above the average for all Aboriginal and/or Torres Strait Islander people living in Greater Melbourne (Northcote/Preston/Whittlesea, Melbourne/Port Phillip, Craigieburn/Sunbury, and Melbourne North-East IAREs).
:::

\newpage

```{r}
# Figure 10G-5
f_caption_figure("(A) Number and (B) age-standardised rate (per 100,000 population) of hospital admissions for ischaemic heart disease for people who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by IARE, 01 July 2017 to 30 June 2021")

figure_number <- hospital_ihd %>%
  dplyr::filter(iare_name %in% iare_names) %>% 
  #
  f_graph_iare(suppress   = "Yes",
               y_variable = hospital_n,
               text_pos   = 2,
               y_label    = hospital_n_label,
               y_max      = NA,
               y_breaks   = 10,
               y_expand   = 5,
               y_title    = "Number of admissions",
               plot_tag   = "A")

figure_rate <- hospital_ihd %>%
  dplyr::filter(iare_name %in% iare_names) %>% 
  #
  f_graph_iare(suppress   = "Yes",
               y_variable = hospital_rate,
               text_pos   = 10,
               y_label    = hospital_rate_label,
               y_max      = NA,
               y_breaks   = 50,
               y_expand   = 35,
               y_title    = "Age-standardised rate",
               plot_tag   = "B")

figure_number + figure_rate
```

#### Data notes:

::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_phidu_vaed)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_iare_nephu)`

`r knitr::asis_output(note_caveat_admissions)`

`r knitr::asis_output(note_caveat_iare_suppress)`

- Admissions for 'ischaemic heart disease' are defined as those with a principal diagnosis code between I20-I25, coded using [ICD-10-AM](https://icd.who.int/browse10/2019/en). The following categories are included:

    - I20: Angina pectoris
    
    - I21: Acute myocardial infarction
    
    - I22: Subsequent myocardial infarction
    
    - I23: Certain current complications following acute myocardial infarction
    
    - I24: Other acute ischaemic heart diseases
    
    - I25: Chronic ischaemic heart disease
:::

\newpage

```{r}
# Table 10G-3
f_caption_table("Hospital admissions for ischaemic heart disease for people who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by IARE, 01 July 2017 to 30 June 2021")

hospital_ihd %>% 
  dplyr::select(iare_name,
                hospital_n_label,
                hospital_rate_label,
                hospital_ratio_label) %>%
  #
  f_table_iare(table_headers = c("Number of\nadmissions", "Age-standardised rate\n(per 100,000)", "Standardised\nratio"),
               last_column   = 4) %>% 
  #
  flextable::width(j = 3,
                   width = 1.75)
```

#### Data notes:

::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_phidu_vaed)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_iare_nephu)`

`r knitr::asis_output(note_caveat_admissions)`

`r knitr::asis_output(note_caveat_iare_ratio)`

`r knitr::asis_output(note_caveat_iare_suppress)`

- Admissions for 'ischaemic heart disease' are defined as those with a principal diagnosis code between I20-I25, coded using [ICD-10-AM](https://icd.who.int/browse10/2019/en). The following categories are included:

    - I20: Angina pectoris
    
    - I21: Acute myocardial infarction
    
    - I22: Subsequent myocardial infarction
    
    - I23: Certain current complications following acute myocardial infarction
    
    - I24: Other acute ischaemic heart diseases
    
    - I25: Chronic ischaemic heart disease
:::

\newpage

## Hospital Admissions - Heart Failure

```{r, include = FALSE}
# ------------------------------------------------------------------------------
# Hospitalisations - heart failure
# Data from PHIDU Social Health Atlas
# ------------------------------------------------------------------------------
# Read in PHIDU data
hospital_failure <- f_read_phidu_iare_hospital(data_sheet    = "Admissions_principal_diag",
                                               column_select = "admissions_for_heart_failure_aboriginal_people")

# Data wrangling
hospital_failure <- hospital_failure %>% 
  dplyr::mutate(iare_name = factor(iare_name, levels = iare_inc_vic_lvl),
                #
                hospital_n_label = round(hospital_n, digits = 0),
                #
                hospital_n_label = dplyr::case_when(
                  is.na(hospital_n) ~ "*",
                  TRUE ~ format(hospital_n_label, big.mark = ",")),
                #
                hospital_rate_label = round(hospital_rate, digits = 0),
                #
                hospital_rate_label = dplyr::case_when(
                  is.na(hospital_rate) ~ "*",
                  TRUE ~ format(hospital_rate_label, big.mark = ",")),
                #
                hospital_ratio_label = round(hospital_ratio, digits = 1),
                #
                hospital_ratio_label = dplyr::case_when(
                  is.na(hospital_ratio) ~ "*",
                  TRUE ~ format(hospital_ratio_label, big.mark = ","))) %>% 
  #
  dplyr::select(iare_name,
                iare_name_short,
                starts_with("hospital")) %>% 
  #
  dplyr::arrange(iare_name)
```


```{r}
#
results <- list()

data <- hospital_failure %>% 
  dplyr::rename(n     = hospital_n,
                rate  = hospital_rate,
                ratio = hospital_ratio)

results$rank_n <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(n)) %>% 
  dplyr::select(iare_name,
                n)

results$rank_rate <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(rate)) %>% 
  dplyr::select(iare_name,
                rate)

results$rank_ratio <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(ratio)) %>% 
  dplyr::select(iare_name,
                ratio)
```

::: {custom-style="Dotpt"}
- Data for this indicator were obtained from the PHIDU Social Health Atlas, compiled from the Victorian Admitted Episodes Dataset.

- During the four-year period 01 July 2017 to 30 June 2021 there were 193 hospital admissions for heart failure for Aboriginal and/or Torres Strait Islander people living in Greater Melbourne.

- The number of admissions for Aboriginal and/or Torres Strait Islander people living in different IAREs within the NEPHU catchment ranged from 26 (Northcote/Preston/Whittlesea IARE) to 5 (Melbourne East IARE).

- The age-standardised rate of admissions for Aboriginal and/or Torres Strait Islander people living in different IAREs within the NEPHU catchment ranged from 399 per 100,000 population (Maroondah IARE) to 59 per 100,000 population (Melbourne North-East IARE).

- 4 of the 8 IAREs in the NEPHU catchment area with data available for reporting had an age-standardised admission rate above the average for all Aboriginal and/or Torres Strait Islander people living in Greater Melbourne (Maroondah, Moreland/Broadmeadows, Craigieburn/Sunbury, and Northcote/Preston/Whittlesea IAREs).
:::

\newpage

```{r}
# Figure 10G-6
f_caption_figure("(A) Number and (B) age-standardised rate (per 100,000 population) of hospital admissions for heart failure for people who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by IARE, 01 July 2017 to 30 June 2021")

figure_number <- hospital_failure %>%
  dplyr::filter(iare_name %in% iare_names) %>% 
  #
  f_graph_iare(suppress   = "Yes",
               y_variable = hospital_n,
               text_pos   = 1,
               y_label    = hospital_n_label,
               y_max      = NA,
               y_breaks   = 5,
               y_expand   = 4.5,
               y_title    = "Number of admissions",
               plot_tag   = "A")

figure_rate <- hospital_failure %>%
  dplyr::filter(iare_name %in% iare_names) %>% 
  #
  f_graph_iare(suppress   = "Yes",
               y_variable = hospital_rate,
               text_pos   = 10,
               y_label    = hospital_rate_label,
               y_max      = NA,
               y_breaks   = 50,
               y_expand   = 7,
               y_title    = "Age-standardised rate",
               plot_tag   = "B")

figure_number + figure_rate
```

#### Data notes:

::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_phidu_vaed)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_iare_nephu)`

`r knitr::asis_output(note_caveat_admissions)`

`r knitr::asis_output(note_caveat_iare_suppress)`

- Admissions for 'heart failure' are defined as those with a principal diagnosis code of I50 (heart failure), coded using [ICD-10-AM](https://icd.who.int/browse10/2019/en).

    - Heart failure following cardiac or obstetric surgery, heart failure due to hypertension, and neonatal cardiac failure are excluded.
:::

\newpage

```{r}
# Table 10G-4
f_caption_table("Hospital admissions for heart failure for people who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by IARE, 01 July 2017 to 30 June 2021")

hospital_failure %>% 
  dplyr::select(iare_name,
                hospital_n_label,
                hospital_rate_label,
                hospital_ratio_label) %>%
  #
  f_table_iare(table_headers = c("Number of\nadmissions", "Age-standardised rate\n(per 100,000)", "Standardised\nratio"),
               last_column   = 4) %>% 
  #
  flextable::width(j = 3,
                   width = 1.75)
```

#### Data notes:

::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_phidu_vaed)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_iare_nephu)`

`r knitr::asis_output(note_caveat_admissions)`

`r knitr::asis_output(note_caveat_iare_ratio)`

`r knitr::asis_output(note_caveat_iare_suppress)`

- Admissions for 'heart failure' are defined as those with a principal diagnosis code of I50 (heart failure), coded using [ICD-10-AM](https://icd.who.int/browse10/2019/en).

    - Heart failure following cardiac or obstetric surgery, heart failure due to hypertension, and neonatal cardiac failure are excluded.
:::

\newpage

# Stroke

## Prevalence

```{r, include = FALSE}
# ------------------------------------------------------------------------------
# Stroke, by LGA
# Data from ABS Census using TableBuilderPro
# ------------------------------------------------------------------------------
# Read in Census data
health_stroke_lga <- f_read_census_lga("Aboriginal_Health_Stroke_LGA.csv",
                                       variable = "stroke") %>% 
  #
  dplyr::filter(stroke != "Not stated")

# Data wrangling
health_stroke_lga <- health_stroke_lga %>%
  #
  dplyr::mutate(stroke = dplyr::case_when(
    stroke == "Has had a stroke" ~ "yes",
    TRUE ~ "no")) %>% 
  #
  f_wrangle_census_lga(variable = stroke,
                       level    = "person",
                       dichot   = "yes")

# ------------------------------------------------------------------------------
# Stroke, by age and sex
# Data from ABS Census using TableBuilderPro
# ------------------------------------------------------------------------------
# Read in Census data
health_stroke_agesex <- f_read_census_agesex("Aboriginal_Health_Stroke_AgeSex.csv",
                                             variable = "stroke") %>% 
  #
  dplyr::filter(stroke != "Not stated")

# Data wrangling
health_stroke_agesex <- health_stroke_agesex %>% 
  #
  dplyr::mutate(stroke = dplyr::case_when(
    stroke == "Has had a stroke" ~ "yes",
    TRUE ~ "no")) %>%   
  #
  f_wrangle_census_agesex(variable = stroke,
                          dichot   = "yes")
```


```{r}
#
results <- list()
data    <- health_stroke_lga

results$rank_n <- data %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  dplyr::arrange(desc(aboriginal_n)) %>% 
  dplyr::select(lga_name,
                aboriginal_n)

results$rank_prop <- data %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  dplyr::arrange(desc(aboriginal_prop)) %>% 
  dplyr::select(lga_name,
                aboriginal_prop)

results$rank_diff <- data %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  dplyr::arrange(desc(aboriginal_difference)) %>% 
  dplyr::select(lga_name,
                aboriginal_difference)
```

::: {custom-style="Dotpt"}
- Data for this indicator were obtained from the ABS Census 2021.

- Overall, 1.2% (n = 131) of people living in the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census have been told by a doctor or nurse that they have had a stroke. This is higher than for non-Indigenous people living in the NEPHU catchment (0.9%).

- The number of people living in different LGAs within the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census who have been told by a doctor or nurse that they have had a stroke ranged from 23 (Darebin and Hume LGAs) to 0 (Manningham and Yarra LGAs).

- The gap between Aboriginal and/or Torres Strait Islander and non-Indigenous people was biggest for Darebin and Hume LGAs.

- The percentage of people living in different LGAs within the NEPHU catchment who identified as Aboriginal and/or Torres Strait Islander in the Census who have been told by a doctor or nurse that they have had a stroke ranged from 2.1% (Boroondara LGA) to 0.0% (Manningham and Yarra LGAs).

- The number of people who have been told by a doctor or nurse that they have had a stroke was highest for females aged 50-59 years.
:::

\newpage

```{r}
# Figure 10G-7
f_caption_figure("(A) Number and (B) percentage of people who identified as Aboriginal and/or Torres Strait Islander and who have been told by a doctor or nurse that they have had a stroke, NEPHU catchment, by LGA, 2021")

figure_number <- health_stroke_lga %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  #
  f_graph_lga(y_variable = aboriginal_n,
              y_max      = NA,
              y_breaks   = 2,
              y_expand   = 1.5,
              y_title    = "Number of people",
              plot_tag   = "A")

figure_percent <- health_stroke_lga %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  #
  f_graph_lga(y_variable = aboriginal_prop,
              y_max      = 5,
              y_breaks   = 0.5,
              y_expand   = 0.125,
              y_title    = "Percentage of people",
              plot_tag   = "B")

figure_number + figure_percent
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`
:::

\newpage

```{r}
# Figure 10G-8
f_caption_figure("Observed vs. expected number of Aboriginal and/or Torres Strait Islander people who have been told by a doctor or nurse that they have had a stroke, NEPHU catchment, by LGA, 2021")

health_stroke_lga %>% 
  dplyr::filter(lga_name %in% nephu_lga) %>% 
  #
  f_graph_expected(variable = lga_name,
                   y_breaks = 2,
                   y_expand = 1.5)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`
:::

\newpage

```{r}
# Table 10G-5
f_caption_table("Observed vs. expected number of Aboriginal and/or Torres Strait Islander people who have been told by a doctor or nurse that they have had a stroke, NEPHU catchment, by LGA, 2021")

health_stroke_lga %>% 
  dplyr::filter(lga_name %in% c(nephu_lga, "NEPHU")) %>%
  #
  f_table_expected(variable       = lga_name,
                   variable_label = "LGA",
                   group_label    = " ",
                   total_row      = 13)
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`
:::

\newpage

```{r}
# Figure 10G-9
f_caption_figure("(A) Number and (B) percentage of people who identified as Aboriginal and/or Torres Strait Islander and who have been told by a doctor or nurse that they have had a stroke, NEPHU catchment, by age and sex, 2021")

figure_female_number <- health_stroke_agesex %>%
  f_graph_agesex(sex        = "Female",
                 y_variable = aboriginal_n,
                 y_max      = 22,
                 y_breaks   = 2,
                 y_expand   = 0.5,
                 y_title    = "Number of females",
                 plot_title = "Females",
                 plot_tag   = "A")

figure_female_percent <- health_stroke_agesex %>%
  f_graph_agesex(sex        = "Female",
                 y_variable = aboriginal_prop,
                 y_max      = 10,
                 y_breaks   = 1,
                 y_expand   = 0.1,
                 y_title    = "Percentage of females",
                 plot_title = " ",
                 plot_tag   = "B")

figure_male_number <- health_stroke_agesex %>%
  f_graph_agesex(sex        = "Male",
                 y_variable = aboriginal_n,
                 y_max      = 22,
                 y_breaks   = 2,
                 y_expand   = 0.5,
                 y_title    = "Number of males",
                 plot_title = "Males",
                 plot_tag   = "A")

figure_male_percent <- health_stroke_agesex %>%
  f_graph_agesex(sex        = "Male",
                 y_variable = aboriginal_prop,
                 y_max      = 10,
                 y_breaks   = 1,
                 y_expand   = 0.1,
                 y_title    = "Percentage of males",
                 plot_title = " ",
                 plot_tag   = "B")

figure_female_number + figure_female_percent
figure_male_number + figure_male_percent
```

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_unknown_agesex)`
:::

\newpage

```{r} 
# Figure 10G-10
f_caption_figure("Observed vs. expected number of Aboriginal and/or Torres Strait Islander people who have been told by a doctor or nurse that they have had a stroke, NEPHU catchment, by age and sex, 2021") 

health_stroke_agesex %>%  
  dplyr::filter(age_group %in% age_group_lvl) %>%  
  # 
  dplyr::mutate(age_group = factor(age_group, levels = age_group_lvl)) %>% 
  #
  f_graph_expected(variable = age_group, 
                   y_breaks = 2, 
                   y_expand = 0.5) +
  #
  facet_wrap(~sex,
             axes = "all")
``` 

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_unknown_agesex)`
:::

\newpage

```{r} 
# Table 10G-6
f_caption_table("Observed vs. expected number of Aboriginal and/or Torres Strait Islander people who have been told by a doctor or nurse that they have had a stroke, NEPHU catchment, by age and sex, 2021") 

health_stroke_agesex %>% 
  dplyr::filter(sex == "Female") %>%  
  # 
  f_table_expected(variable       = age_group,
                   variable_label = "Age group",
                   group_label    = "Females",
                   total_row      = 12)
  
health_stroke_agesex %>% 
  dplyr::filter(sex == "Male") %>%  
  # 
  f_table_expected(variable       = age_group,
                   variable_label = "Age group",
                   group_label    = "Males",
                   total_row      = 12)
``` 

#### Data notes:
 
::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_census)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_unknown_agesex)`
:::

\newpage

# Circulatory System Diseases

## Hospital Admissions

```{r, include = FALSE}
# ------------------------------------------------------------------------------
# Hospitalisations - circulatory system diseases
# Data from PHIDU Social Health Atlas
# ------------------------------------------------------------------------------
# Read in PHIDU data
hospital_circ <- f_read_phidu_iare_hospital(data_sheet    = "Admissions_principal_diag",
                                            column_select = "admissions_for_circulatory_system_diseases_aboriginal_people")

# Data wrangling
hospital_circ <- hospital_circ %>% 
  dplyr::mutate(iare_name = factor(iare_name, levels = iare_inc_vic_lvl)) %>% 
  #
  dplyr::select(iare_name,
                iare_name_short,
                starts_with("hospital")) %>% 
  #
  dplyr::arrange(iare_name)
```


```{r}
#
results <- list()

data <- hospital_circ %>% 
  dplyr::rename(n     = hospital_n,
                rate  = hospital_rate,
                ratio = hospital_ratio)

results$rank_n <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(n)) %>% 
  dplyr::select(iare_name,
                n)

results$rank_rate <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(rate)) %>% 
  dplyr::select(iare_name,
                rate)

results$rank_ratio <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(ratio)) %>% 
  dplyr::select(iare_name,
                ratio)
```

::: {custom-style="Dotpt"}
- Data for this indicator were obtained from the PHIDU Social Health Atlas, compiled from the Victorian Admitted Episodes Dataset.

- During the four-year period 01 July 2017 to 30 June 2021 there were 1,868 hospital admissions for circulatory system diseases for Aboriginal and/or Torres Strait Islander people living in Greater Melbourne.

- The number of admissions for Aboriginal and/or Torres Strait Islander people living in different IAREs within the NEPHU catchment ranged from 284 (Northcote/Preston/Whittlesea IARE) to 26 (Whitehorse IARE).

- The age-standardised rate of admissions for Aboriginal and/or Torres Strait Islander people living in different IAREs within the NEPHU catchment ranged from 1,739 per 100,000 population (Northcote/Preston/Whittlesea IARE) to 738 per 100,000 population (Yarra Ranges IARE).

- 6 of the 10 IAREs in the NEPHU catchment area had an age-standardised admission rate above the average for all Aboriginal and/or Torres Strait Islander people living in Greater Melbourne (Northcote/Preston/Whittlesea, Melbourne/Port Phillip, Craigieburn/Sunbury, Maroondah, Melbourne North-East, and Moreland/Broadmeadows IAREs).
:::

\newpage

```{r}
# Figure 10G-11
f_caption_figure("(A) Number and (B) age-standardised rate (per 100,000 population) of hospital admissions for circulatory system diseases for people who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by IARE, 01 July 2017 to 30 June 2021")

figure_number <- hospital_circ %>%
  dplyr::filter(iare_name %in% iare_names) %>% 
  #
  f_graph_iare(suppress   = "No",
               y_variable = hospital_n,
               y_max      = NA,
               y_breaks   = 50,
               y_expand   = 20,
               y_title    = "Number of admissions",
               plot_tag   = "A")

figure_rate <- hospital_circ %>%
  dplyr::filter(iare_name %in% iare_names) %>% 
  #
  f_graph_iare(suppress   = "No",
               y_variable = hospital_rate,
               y_max      = NA,
               y_breaks   = 200,
               y_expand   = 75,
               y_title    = "Age-standardised rate",
               plot_tag   = "B")

figure_number + figure_rate
```

#### Data notes:

::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_phidu_vaed)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_iare_nephu)`

`r knitr::asis_output(note_caveat_admissions)`

- Admissions for 'circulatory system diseases' are defined as those with a principal diagnosis code between I00-I99, coded using [ICD-10-AM](https://icd.who.int/browse10/2019/en). The following major categories are included:

    - I00-I02: Acute rheumatic fever
    
    - I05-I09: Chronic rheumatic heart diseases
    
    - I10-I15: Hypertensive diseases
    
    - I20-I25: Ischaemic heart diseases
    
    - I26-I28: Pulmonary heart disease and diseases of pulmonary circulation
    
    - I30-I52: Other forms of heart disease
    
    - I60-I69: Cerebrovascular diseases
    
    - I70-I79: Diseases of arteries, arterioles, and capillaries
    
    - I80-I89: Diseases of veins, lymphatic vessels, and lymph nodes, not elsewhere classified
    
    - I95-I99: Other and unspecified disorders of the circulatory system
:::

\newpage

```{r}
# Table 10G-7
f_caption_table("Hospital admissions for circulatory system diseases for people who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by IARE, 01 July 2017 to 30 June 2021")

hospital_circ %>% 
  dplyr::select(-iare_name_short) %>% 
  #
  dplyr::mutate(hospital_rate  = round(hospital_rate, digits = 0),
                hospital_ratio = round(hospital_ratio, digits = 1)) %>% 
  #
  f_table_iare(table_headers = c("Number of\nadmissions", "Age-standardised rate\n(per 100,000)", "Standardised\nratio"),
               last_column   = 4) %>% 
  #
  flextable::width(j = 3,
                   width = 1.75)
```

#### Data notes:

::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_phidu_vaed)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_iare_nephu)`

`r knitr::asis_output(note_caveat_admissions)`

`r knitr::asis_output(note_caveat_iare_ratio)`

- Admissions for 'circulatory system diseases' are defined as those with a principal diagnosis code between I00-I99, coded using [ICD-10-AM](https://icd.who.int/browse10/2019/en). The following major categories are included:

    - I00-I02: Acute rheumatic fever
    
    - I05-I09: Chronic rheumatic heart diseases
    
    - I10-I15: Hypertensive diseases
    
    - I20-I25: Ischaemic heart diseases
    
    - I26-I28: Pulmonary heart disease and diseases of pulmonary circulation
    
    - I30-I52: Other forms of heart disease
    
    - I60-I69: Cerebrovascular diseases
    
    - I70-I79: Diseases of arteries, arterioles, and capillaries
    
    - I80-I89: Diseases of veins, lymphatic vessels, and lymph nodes, not elsewhere classified
    
    - I95-I99: Other and unspecified disorders of the circulatory system
:::

\newpage

## Emergency Presentations

```{r, include = FALSE}
# ------------------------------------------------------------------------------
# Emergency presentations - circulatory system diseases
# Data from PHIDU Social Health Atlas
# ------------------------------------------------------------------------------
# Read in PHIDU data
emergency_circ <- f_read_phidu_iare_hospital(data_sheet    = "ED_total",
                                             column_select = "emergency_department_presentations_total_presentations_for_diseases_of_the_circulatory_system_aboriginal_people")

# Data wrangling
emergency_circ <- emergency_circ %>% 
  dplyr::mutate(iare_name = factor(iare_name, levels = iare_inc_vic_lvl)) %>% 
  #
  dplyr::arrange(iare_name)
```


```{r}
#
results <- list()

data <- emergency_circ %>% 
  dplyr::rename(n     = hospital_n,
                rate  = hospital_rate,
                ratio = hospital_ratio)

results$rank_n <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(n)) %>% 
  dplyr::select(iare_name,
                n)

results$rank_rate <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(rate)) %>% 
  dplyr::select(iare_name,
                rate)

results$rank_ratio <- data %>% 
  dplyr::filter(iare_name %in% iare_names) %>% 
  dplyr::arrange(desc(ratio)) %>% 
  dplyr::select(iare_name,
                ratio)
```

::: {custom-style="Dotpt"}
- Data for this indicator were obtained from the PHIDU Social Health Atlas, compiled from the AIHW National Non-Admitted Patient Emergency Department Care Database.

- Between 01 July 2019 and 30 June 2020 there were 363 emergency department presentations for circulatory system diseases for Aboriginal and/or Torres Strait Islander people living in Greater Melbourne.

- The number of presentations for Aboriginal and/or Torres Strait Islander people living in different IAREs within the NEPHU catchment ranged from 45 (Northcote/Preston/Whittlesea IARE) to 7 (Knox and Maroondah IAREs).

- The age-standardised rate of presentations for Aboriginal and/or Torres Strait Islander people living in different IAREs within the NEPHU catchment ranged from 1,665 per 100,000 population (Yarra Ranges IARE) to 640 per 100,000 population (Melbourne East IARE).

- 6 of the 10 IAREs in the NEPHU catchment area had an age-standardised presentation rate above the average for all Aboriginal and/or Torres Strait Islander people living in Greater Melbourne (Yarra Ranges, Melbourne North-East, Moreland/Broadmeadows, Northcote/Preston/Whittlesea, Melbourne/Port Phillip, and Whitehorse IAREs).
:::

\newpage

```{r}
# Figure 10G-12
f_caption_figure("(A) Number and (B) age-standardised rate (per 100,000 population) of emergency department presentations for circulatory system diseases for people who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by IARE, 01 July 2019 to 30 June 2020")

figure_number <- emergency_circ %>%
  dplyr::filter(iare_name %in% iare_names) %>% 
  #
  f_graph_iare(suppress   = "No",
               y_variable = hospital_n,
               y_max      = NA,
               y_breaks   = 5,
               y_expand   = 1,
               y_title    = "Number of presentations",
               plot_tag   = "A")

figure_rate <- emergency_circ %>%
  dplyr::filter(iare_name %in% iare_names) %>% 
  #
  f_graph_iare(suppress   = "No",
               y_variable = hospital_rate,
               y_max      = NA,
               y_breaks   = 200,
               y_expand   = 150,
               y_title    = "Age-standardised rate",
               plot_tag   = "B")

figure_number + figure_rate
```

#### Data notes:

::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_phidu_vemd)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_iare_nephu)`

- Presentations for 'circulatory system diseases' are defined as those with a principal diagnosis code between I00-I99, coded using [ICD-10-AM](https://icd.who.int/browse10/2019/en). The following major categories are included:

    - I00-I02: Acute rheumatic fever
    
    - I05-I09: Chronic rheumatic heart diseases
    
    - I10-I15: Hypertensive diseases
    
    - I20-I25: Ischaemic heart diseases
    
    - I26-I28: Pulmonary heart disease and diseases of pulmonary circulation
    
    - I30-I52: Other forms of heart disease
    
    - I60-I69: Cerebrovascular diseases
    
    - I70-I79: Diseases of arteries, arterioles, and capillaries
    
    - I80-I89: Diseases of veins, lymphatic vessels, and lymph nodes, not elsewhere classified
    
    - I95-I99: Other and unspecified disorders of the circulatory system
:::

\newpage

```{r}
# Table 10G-8
f_caption_table("Emergency department presentations for circulatory system diseases for people who identified as Aboriginal and/or Torres Strait Islander, NEPHU catchment, by IARE, 01 July 2019 to 30 June 2020")

emergency_circ %>% 
  dplyr::select(-iare_name_short) %>% 
  #
  dplyr::mutate(hospital_rate  = round(hospital_rate, digits = 0),
                hospital_ratio = round(hospital_ratio, digits = 1)) %>% 
  #
  f_table_iare(table_headers = c("Number of\npresentations", "Age-standardised rate\n(per 100,000)", "Standardised\nratio"),
               last_column   = 4) %>% 
  #
  flextable::width(j = 3,
                   width = 1.75)
```

#### Data notes:

::: {custom-style="Footnotes"}
`r knitr::asis_output(note_data_phidu_vemd)`

`r knitr::asis_output(note_caveat_underidentification)`

`r knitr::asis_output(note_caveat_iare_nephu)`

`r knitr::asis_output(note_caveat_iare_ratio)`

- Presentations for 'circulatory system diseases' are defined as those with a principal diagnosis code between I00-I99, coded using [ICD-10-AM](https://icd.who.int/browse10/2019/en). The following major categories are included:

    - I00-I02: Acute rheumatic fever
    
    - I05-I09: Chronic rheumatic heart diseases
    
    - I10-I15: Hypertensive diseases
    
    - I20-I25: Ischaemic heart diseases
    
    - I26-I28: Pulmonary heart disease and diseases of pulmonary circulation
    
    - I30-I52: Other forms of heart disease
    
    - I60-I69: Cerebrovascular diseases
    
    - I70-I79: Diseases of arteries, arterioles, and capillaries
    
    - I80-I89: Diseases of veins, lymphatic vessels, and lymph nodes, not elsewhere classified
    
    - I95-I99: Other and unspecified disorders of the circulatory system
:::
